chore: move some js out of image php class
