+++ admin/controllers/images/views/show/style.css
@@ -1,3 +1,25 @@
+#upload_space {
+	height:10rem;
+	padding:1rem;
+	margin:1rem;
+	border:2px dashed #aaa;
+	display:flex;
+	align-items: center;
+	justify-content: center;
+	transition:all 0.3s ease;
+	
+	h1 {
+		font-size:2rem;
+		opacity:0.3;
+		font-weight:900;
+	}
+	
+	&.ready {
+		border:2px dashed #aaa;
+		background:#cec;
+	}
+}
+
 #all_images {
 	display:flex;
 	flex-wrap:wrap;
